auth_service/requirements.txt
# auth_service/requirements.txt


Django==5.2
dj-database-url==1.0.0
python-dotenv==1.0.1
djangorestframework
drf-spectacular
drf-spectacular[sidecar]
djangorestframework-simplejwt==5.5.0
celery
redis
